requirements.txt
# File: Portfolio_Tracker-main/requirements.txt
# Description: List of dependencies required for the project.

Flask==2.0.3
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
Flask-Login==0.5.0
Werkzeug==2.0.3
requests==2.26.0
email-validator==1.1.3
WTForms==3.0.1
python-dotenv==0.19.2
SQLAlchemy==1.4.32
flask_migrate==4.0.4
coverage==7.3.1
pytest==7.4.0
pytest-flask==1.2.0
factory_boy==3.0.0
Flask-Testing==0.8.1
